refactor(structure): reorganiza linear_search para estrutura de gem
