+++ docs/Libear_search.md
