+++ lib/ruby_coding_challenges/linear_search/implementation_each_with_index.rb
@@ -1,0 +1,15 @@
+# Linear search (busca linear) - Mais informações readme
+
+module RubyCodingChallenges
+    module LinearSearch
+      module EachWithIndex
+        # Versão usando .each_with_index (mais idiomática, dispensa variável de índice manual)
+        def self.search(array, target)
+          array.each_with_index do |amount, index|
+            return index if amount == target
+          end
+          return -1
+        end
+      end
+    end
+end
